generated_code/arcgis-sys-api-dummy-flask/requirements.txt
Flask>=2.3.0
Flask-SQLAlchemy>=3.0.0
marshmallow>=3.19.0
flask-marshmallow>=0.15.0
marshmallow-sqlalchemy>=0.29.0
Flask-JWT-Extended>=4.5.0
Flask-CORS>=4.0.0
Werkzeug>=2.3.0
